Change usage of float to double.

+++ src/main/java/server/ServerController.java
@@ -226,7 +226,7 @@ public class ServerController {
             (int) Math.round(paperRecyclingRequest.getAmount()));
         double co2 = DatabaseHandler.increaseCO2RedBy(paperRecyclingRequest.getUid(),
             paperRecyclingRequest.getAmount() * 1.21);
-        float amount = DatabaseHandler.increaseFeatureCounter(paperRecyclingRequest.getUid(),
+        double amount = DatabaseHandler.increaseFeatureCounter(paperRecyclingRequest.getUid(),
             "paperrecycling",
             paperRecyclingRequest.getAmount());
